chore: remove dashboardQueryRangeCalled from global dashboard state (#10650)
* chore: remove dashboardQueryRangeCalled from global dashboard state

* chore: cleanup dashboard page setup (#10652)
